Add basic Messaging interface, which allows for generalized storage and editing of in-game text.
It also includes a Command interface for command names, so those can
be edited as well- but I'm not entirely sure how/if this should fit
into the existing Bukkit command system.

I've started to use it a bit internally within Persistence as a
proof-of-concept, but I've yet to convert all of the text over.