Change the way permissions work by default, so that access is allowed if Permissions is not present. This assumes the admin is OK with a wide-open server, or that you're doing your own auth.
Added a new PermissionsType, ADMIN_ONLY - which is what Persistence
now uses. This restores the old "deny by default" behavior, which is
what Persistence wants, but may not be what the average plugin wants.

Also add versions of getCommand that take a permissions type but not a
node name override, since it seems like messing with the type is
something you may want to do.